requirements.txt
pulsar-client==2.10.2
viaa-chassis==0.2.0rc1
meemoo-cloudevents==0.1.0rc3